Handle test case coverage issue
