+++ app/client/src/components/designSystems/appsmith/TableComponent/TableDataDownload.tsx
@@ -5,16 +5,44 @@ import { ReactComponent as DownloadIcon } from "assets/icons/control/download-ta
 import { ReactTableColumnProps } from "components/designSystems/appsmith/TableComponent/Constants";
 import { TableIconWrapper } from "components/designSystems/appsmith/TableComponent/TableStyledWrappers";
 import TableActionIcon from "components/designSystems/appsmith/TableComponent/TableActionIcon";
-import {
-  downloadTableDataAsCSV,
-  transformTableDataIntoCsv,
-} from "./CommonUtilities";
+import { transformTableDataIntoCsv } from "./CommonUtilities";
 
 interface TableDataDownloadProps {
   data: Array<Record<string, unknown>>;
   columns: ReactTableColumnProps[];
   widgetName: string;
 }
+
+const downloadDataAsCSV = (props: {
+  csvData: Array<Array<any>>;
+  fileName: string;
+}) => {
+  let csvContent = "";
+  props.csvData.forEach((infoArray: Array<any>, index: number) => {
+    const dataString = infoArray.join(",");
+    csvContent += index < props.csvData.length ? dataString + "\n" : dataString;
+  });
+  const anchor = document.createElement("a");
+  const mimeType = "application/octet-stream";
+  if (navigator.msSaveBlob) {
+    navigator.msSaveBlob(
+      new Blob([csvContent], {
+        type: mimeType,
+      }),
+      props.fileName,
+    );
+  } else if (URL && "download" in anchor) {
+    anchor.href = URL.createObjectURL(
+      new Blob([csvContent], {
+        type: mimeType,
+      }),
+    );
+    anchor.setAttribute("download", props.fileName);
+    document.body.appendChild(anchor);
+    anchor.click();
+    document.body.removeChild(anchor);
+  }
+};
 
 const TableDataDownload = (props: TableDataDownloadProps) => {
   const [selected, toggleButtonClick] = React.useState(false);
@@ -24,7 +52,7 @@ const TableDataDownload = (props: TableDataDownloadProps) => {
       columns: props.columns,
       data: props.data,
     });
-    downloadTableDataAsCSV({
+    downloadDataAsCSV({
       csvData: csvData,
       fileName: `${props.widgetName}.csv`,
     });
